Update RecipientType logic in CustomTechModel
Refactor the logic for setting `RecipientType` in the
`CustomTechModel` class. When `IsOnDemand` is true and
`RecipientType` is `Human`, it is changed to
`GlobalObtainable`. A new condition ensures that if
`VillagersToSpawn` is greater than 0 and
`RecipientType` is `GlobalObtainable`, it reverts to
`Human`. This change allows the random map script to
properly disable a tech after it has been received.